feat: filter transactions from the mobile filter panel
